Refactor screens auth, deleted sign-up module and replace it with auth module
