+++ src/app/app-routing.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
-  { path: '', redirectTo: 'sign-up', pathMatch: 'full' },
+  { path: '', redirectTo: 'auth', pathMatch: 'full' },
   {
-    path: 'sign-up',
-    loadChildren: () => import('./screens/sign-up/sign-up.module').then(m => m.SignUpModule)
+    path: 'auth',
+    loadChildren: () => import('./screens/auth/auth.module').then(m => m.AuthModule)
   }
 ];
