+++ src/app/screens/auth/auth-routing.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { SignUpComponent } from './sign-up.component';
+import { SignUpComponent } from './sign-up/sign-up.component';
 
 const routes: Routes = [
-  { path: '', component: SignUpComponent }
+  { path: '', pathMatch: 'full', redirectTo: 'sign-up' },
+  { path: 'sign-up', component: SignUpComponent },
 ];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class SignUpRoutingModule { }
+export class AuthRoutingModule { }
