+++ src/app/screens/auth/sign-up/sign-up.component.scss
