+++ src/app/screens/auth/sign-up/sign-up.component.ts
@@ -65,7 +65,7 @@ export class SignUpComponent implements OnInit {
   signUp(): void {
     this.service.createUser(this.form.value).subscribe(() => {
       console.log('success');
-      this.route.navigate(['/login'])
+      this.route.navigate(['../login'])
     }, error => {
       console.error(error.error.errors.full_messages.join());
     });
